Tests/LibWeb/Text/expected/textarea-clear.txt

1. value after clear: ""
2. defaultValue after clear: "initial content"
